Backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
marshmallow
psycopg2-binary
python-dotenv
werkzeug
email-validator